Move average_grads_batch_torch to util
