+++ dppum/util.py
@@ -321,3 +321,46 @@ def print_dictionary(dictionary):
     for key, value in dictionary.items():
         print(f"{key}: {value}")
     print("}")
+    
+
+
+def average_grads_batch_torch(tensor_list):
+    """
+    Compute the average of tensors over the batch dimension. This is designed
+    to average a list of lists of gradients into one single list of gradients.
+
+    Parameters
+    ----------
+    tensor_list : list of list of torch.Tensor
+        The input list of lists of tensors. The dimensions of this list of
+        lists are [batch,num_tensors,tensor_sizes].
+
+    Returns
+    -------
+    list of torch.Tensor
+        The output list of averaged tensors. The dimensions of this list are
+        [num_tensors,tensor_sizes].
+
+    """
+    # Initialize a list to store the averaged tensors
+    averaged_tensors = []
+
+    # Iterate over the number of tensors
+    for i in range(len(tensor_list[0])):
+        # Initialize a list to store the tensors of the current index from each batch
+        tensors = []
+
+        # Iterate over the batches
+        for j in range(len(tensor_list)):
+            # Append the tensor of the current index from the current batch to the list
+            tensors.append(tensor_list[j][i])
+
+        # Convert the list of tensors to a 3D tensor
+        tensors = torch.stack(tensors)
+
+        # Calculate the mean over the batch dimension (0th dimension) and
+        # append the result to the averaged tensors list
+        averaged_tensors.append(torch.mean(tensors, dim=0))
+
+    # Return the list of averaged tensors
+    return averaged_tensors
